Rename `useOptionChronogramTask.ts` to `useOptionChronogram.ts`
